refactor: unify genro_builders.builders → genro_builders.builder
- Replace all imports from genro_builders.builders (alias) with
  genro_builders.builder (canonical package) across tests and examples.
- Remove the genro_builders/builders/ alias package entirely.
- Fix unused import (parse_pointer) in _reactivity_mixin.py.

608 tests pass, ruff clean, coverage 91%.
