+++ tests/test_builders/test_annotated_validation.py
@@ -10,7 +10,7 @@ import pytest
 from genro_builders import BagBuilderBase
 from genro_builders.builder import SchemaBuilder
 from genro_builders.builder_bag import BuilderBag as Bag
-from genro_builders.builders import Range, Regex, element
+from genro_builders.builder import Range, Regex, element
 
 
 # =============================================================================
