+++ tests/test_builders/test_decorators.py
@@ -5,7 +5,7 @@ import pytest
 
 from genro_builders import BagBuilderBase
 from genro_builders.builder_bag import BuilderBag as Bag
-from genro_builders.builders import abstract, component, element
+from genro_builders.builder import abstract, component, element
 
 
 # =============================================================================
