+++ tests/test_compile_parent.py
@@ -3,7 +3,7 @@
 
 from genro_builders import BagBuilderBase, BagCompilerBase
 from genro_builders.builder_bag import BuilderBag as Bag
-from genro_builders.builders import element
+from genro_builders.builder import element
 from genro_builders.compiler import compiler
