+++ tests/test_reactivity_edge_cases.py
@@ -7,7 +7,7 @@ import pytest
 
 from genro_builders import BagBuilderBase
 from genro_builders.builder_bag import BuilderBag as Bag
-from genro_builders.builders import element
+from genro_builders.builder import element
 from genro_builders.contrib.html import HtmlBuilder
