+++ tests/test_yaml_renderer.py
@@ -5,7 +5,7 @@ import yaml
 
 from genro_builders import BagBuilderBase
 from genro_builders.builder_bag import BuilderBag as Bag
-from genro_builders.builders import element
+from genro_builders.builder import element
 from genro_builders.compilers import YamlRendererBase
